Refactor - move RescueHandler to own file; create Index struct
