+++ lib/rescue_unique_constraint/adapter/postgresql_adapter.rb
@@ -2,7 +2,7 @@ module RescueUniqueConstraint
   module Adapter
     class PostgresqlAdapter
       def index_error?(index, error_message)
-        error_message[/#{index[0]}/]
+        error_message[/#{index.name}/]
       end
     end
   end
