+++ lib/rescue_unique_constraint/adapter/sqlite_adapter.rb
@@ -2,7 +2,7 @@ module RescueUniqueConstraint
   module Adapter
     class SqliteAdapter
       def index_error?(index, error_message)
-        error_message[/UNIQUE.*#{index[1]}/]
+        error_message[/UNIQUE.*#{index.field}/]
       end
     end
   end
