+++ lib/rescue_unique_constraint/index.rb
@@ -1,0 +1,9 @@
+module RescueUniqueConstraint
+  class Index
+    attr_reader :name, :field
+    def initialize(name, field)
+      @name = name
+      @field = field
+    end
+  end
+end
